pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "connect4-zero"
version = "0.1.0"
description = "AlphaZero-style learner for Connect 4"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
c4z = "connect4zero.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/connect4zero"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
